correct packages data-module
