+++ data/src/main/java/com/example/compmoviedb/data/mappers/moviespopular/ListMoviesPopularEntityMapper.kt
@@ -1,7 +1,7 @@
 package com.example.compmoviedb.data.mappers.moviespopular
 
 import com.example.compmoviedb.data.mappers.MapperData
-import com.example.compmoviedb.data.dataSource.models.moviespopular.ListMoviesPopularEntity
+import com.example.compmoviedb.data.models.moviespopular.ListMoviesPopularEntity
 import com.example.compmoviedb.domain.models.moviespopular.ListMoviesPopularD
 
 class ListMoviesPopularEntityMapper : MapperData<ListMoviesPopularEntity, ListMoviesPopularD> {
