+++ data/src/main/java/com/example/compmoviedb/data/models/moviedetails/ProductionCountry.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.dataSource.models.moviedetails
+package com.example.compmoviedb.data.models.moviedetails
 
 data class ProductionCountry(
     val iso_3166_1: String?,
