+++ data/src/main/java/com/example/compmoviedb/data/models/moviespopular/ListMoviesPopularEntity.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.dataSource.models.moviespopular
+package com.example.compmoviedb.data.models.moviespopular
 
 data class ListMoviesPopularEntity(
     val page: Int,
