+++ data/src/main/java/com/example/compmoviedb/data/remote/dataSource/themoviedb/MovieDBConstants.kt
@@ -1,4 +1,4 @@
-package com.example.compmoviedb.data.dataSource.themoviedb
+package com.example.compmoviedb.data.remote.dataSource.themoviedb
 
 object MovieDBConstants {
     const val MOVIE_DB_KEY = "e57c71791b41b1b6d48678746aa69e44"
