+++ data/src/main/java/com/example/compmoviedb/data/remote/dataSource/themoviedb/TheMovieDBMoviesDataSource.kt
@@ -1,10 +1,10 @@
-package com.example.compmoviedb.data.dataSource.themoviedb
+package com.example.compmoviedb.data.remote.dataSource.themoviedb
 
-import com.example.compmoviedb.data.dataSource.MoviesDataSource
-import com.example.compmoviedb.data.dataSource.models.movieactors.ListActorsMovieEntity
-import com.example.compmoviedb.data.dataSource.models.moviedetails.MovieDetailsEntity
-import com.example.compmoviedb.data.dataSource.models.moviespopular.ListMoviesPopularEntity
-import com.example.compmoviedb.data.dataSource.models.movievideo.MovieVideoEntity
+import com.example.compmoviedb.data.remote.dataSource.MoviesDataSource
+import com.example.compmoviedb.data.models.movieactors.ListActorsMovieEntity
+import com.example.compmoviedb.data.models.moviedetails.MovieDetailsEntity
+import com.example.compmoviedb.data.models.moviespopular.ListMoviesPopularEntity
+import com.example.compmoviedb.data.models.movievideo.MovieVideoEntity
 import com.example.compmoviedb.domain.models.Response
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.channels.awaitClose
